requirements.txt
# YOLOv5 requirements
# Usage: pip install -r requirements.txt

# Python versiyonu
python>=3.8  # Python sürümü minimum 3.8 olmalıdır

# CUDA ve GPU Gereksinimleri --------------------------------------------------
# CUDA: 11.1 veya daha üstü (gerekiyorsa)
# NVIDIA GPU: Tesla, T4, V100, A100 veya benzeri GPU'lar önerilir

# Base ------------------------------------------------------------------------
gitpython>=3.1.30
matplotlib>=3.3
numpy>=1.23.5
opencv-python>=4.1.1
pillow>=10.3.0
psutil  # system resources
PyYAML>=5.3.1
requests>=2.32.2
scipy>=1.4.1
thop>=0.1.1  # FLOPs computation
torch>=1.8.0  # see https://pytorch.org/get-started/locally (recommended)
torchvision>=0.9.0
tqdm>=4.66.3
ultralytics>=8.2.34  # https://ultralytics.com
protobuf<=3.20.1  # See YOLOv5 compatibility issues

# Logging ---------------------------------------------------------------------
tensorboard>=2.4.1  # Tensorboard logging
# clearml>=1.2.0  # ClearML logging (optional)
# comet  # Comet ML logging (optional)

# Plotting --------------------------------------------------------------------
pandas>=1.1.4  # For data handling and analysis
seaborn>=0.11.0  # For advanced plotting

# Export ----------------------------------------------------------------------
coremltools>=6.0  # For CoreML model export
onnx>=1.10.0  # ONNX model export
onnx-simplifier>=0.4.1  # Simplify ONNX model
nvidia-pyindex  # TensorRT model export
nvidia-tensorrt  # TensorRT model export
scikit-learn<=1.1.2  # For CoreML quantization and export
tensorflow>=2.4.0,<=2.13.1  # TensorFlow model export
tensorflowjs>=3.9.0  # TF.js export for web applications
openvino-dev>=2023.0  # OpenVINO model export for Intel hardware

# Deploy ----------------------------------------------------------------------
setuptools>=70.0.0  # For deployment and environment setup
tritonclient[all]~=2.24.0  # NVIDIA Triton inference server (optional)

# Extras ----------------------------------------------------------------------
ipython  # For interactive notebooks
mss  # For screenshots
albumentations>=1.0.3  # For advanced image augmentations
pycocotools>=2.0.6  # For COCO evaluation metrics